@@ -101,28 +101,30 @@
     "vary": "^1.1.2"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "npm run build && npm run jasmine",
     "build": "npx tsc",
     "start": "nodemon src/index.ts",
     "jasmine": "jasmine",
-    "lint": "eslint 'src'",
+    "lint": "eslint src/",
     "prettier": "prettier --config .prettierrc . --write"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/body-parser": "^1.19.5",
+    "@eslint/js": "^9.13.0",
     "@types/express": "^5.0.0",
     "@types/jasmine": "^5.1.4",
-    "@types/node": "^22.7.4",
-    "@types/sharp": "^0.32.0",
     "@types/supertest": "^6.0.2",
-    "eslint": "^8.8.0",
+    "@typescript-eslint/eslint-plugin": "^8.10.0",
+    "@typescript-eslint/parser": "^8.10.0",
+    "eslint": "^8.57.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "globals": "^15.11.0",
     "module-name": "^0.0.1-security",
     "nodemon": "^3.1.7",
     "prettier": "^2.5.1",
-    "typescript": "^5.6.3"
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.10.0"
   }
 }
