@@ -1,21 +1,29 @@
-import express from 'express';
-import sharp from 'sharp';
-const fs = require('fs');
-const path = require('path');
-const outPut = path.resolve('output');
-function imageProcess(filename: string, widthImage: string, heightImage: string) {
-    const outPutFile = path.join(outPut, `${filename}-${widthImage}-${widthImage}.jpg`);
-    const imagePath = path.join(__dirname, filename + '.jpg');
-    if (fs.existsSync(imagePath)) {
-        var width = parseInt(widthImage as string, 10) || 300;
-        var height = parseInt(heightImage as string, 10) || 200;
+import sharp from 'sharp'
+import fs from 'fs'
+import path from 'path'
+const outPutPath = path.resolve('output')
+async function imageProcess(
+  filename: string,
+  widthImage: string,
+  heightImage: string,
+) {
+  const outImagePath = path.join(
+    outPutPath,
+    `${filename}-${widthImage}-${widthImage}.jpg`,
+  )
+  const inputImagePath = path.join(__dirname, filename + '.jpg')
+  //check image exist
+  if (fs.existsSync(outImagePath)) {
+    return outImagePath
+  }
+  if (fs.existsSync(inputImagePath)) {
+    const width = parseInt(widthImage as string, 10) || 300
+    const height = parseInt(heightImage as string, 10) || 200
 
-        sharp(imagePath)
-            .resize(width, height)
-            .toFile(outPutFile);
-        return outPutFile;
-    } else {
-        return null;
-    }
+    await sharp(inputImagePath).resize(width, height).toFile(outImagePath)
+    return outImagePath
+  } else {
+    return null
+  }
 }
-export default imageProcess;
+export default imageProcess
